+++ src/redturtle/prenotazioni/restapi/serializers/adapters/prenotazione.py
@@ -100,6 +100,7 @@ class PrenotazioneSerializer:
             "requirements": requirements,
             "modification_date": json_compatible(self.prenotazione.modified()),
             "creation_date": json_compatible(self.prenotazione.created()),
+            "additional_fields": self.prenotazione.additional_fields,
         }