+++ src/redturtle/prenotazioni/tests/test_add_booking.py
@@ -97,7 +97,7 @@ class TestBookingRestAPIAdd(unittest.TestCase):
             week_table=WEEK_TABLE_SCHEMA,
         )
 
-        booking_type_A = api.content.create(
+        self.booking_type_A = api.content.create(
             type="PrenotazioneType",
             title="Type A",
             duration=30,
@@ -106,7 +106,7 @@ class TestBookingRestAPIAdd(unittest.TestCase):
         )
 
         api.content.transition(obj=self.folder_prenotazioni, transition="publish")
-        api.content.transition(obj=booking_type_A, transition="publish")
+        api.content.transition(obj=self.booking_type_A, transition="publish")
 
         transaction.commit()
 
@@ -374,6 +374,101 @@ class TestBookingRestAPIAdd(unittest.TestCase):
         )
         self.assertEqual(res.status_code, 200)
         self.assertEqual(res.json()["booking_status"], "pending")
+
+    def test_addtional_fields_text(self):
+        # Field tyes for the type field are defined in the redturtle.prenotazioni.booking_additional_fields_types vocabulary
+        self.booking_type_A.booking_additional_fields_schema = [
+            {
+                "name": "text line",
+                "description": "text field description",
+                "type": "textline",
+            }
+        ]
+
+        transaction.commit()
+
+        self.api_session.auth = None
+
+        # Test positive
+        res = self.api_session.post(
+            self.folder_prenotazioni.absolute_url() + "/@booking",
+            json={
+                "booking_date": "%sT09:00:00"
+                % (date.today() + timedelta(1)).strftime("%Y-%m-%d"),
+                "booking_type": "Type A",
+                "fields": [
+                    {"name": "title", "value": "Mario Rossi"},
+                    {"name": "email", "value": "mario.rossi@example"},
+                ],
+                "gate": "Gate A",
+                "additional_fields": [
+                    {"name": "text line", "value": "text field value"}
+                ],
+            },
+        )
+        self.assertEqual(res.status_code, 200)
+        self.assertEqual(
+            res.json()["additional_fields"],
+            '[{"name": "text line", "value": "text field value"}]',
+        )
+
+        # Text wrong field type
+        res = self.api_session.post(
+            self.folder_prenotazioni.absolute_url() + "/@booking",
+            json={
+                "booking_date": "%sT09:00:00"
+                % (date.today() + timedelta(2)).strftime("%Y-%m-%d"),
+                "booking_type": "Type A",
+                "fields": [
+                    {"name": "title", "value": "Mario Rossi"},
+                    {"name": "email", "value": "mario.rossi@example"},
+                ],
+                "gate": "Gate A",
+                "additional_fields": [{"name": "text line", "value": 2}],
+            },
+        )
+        self.assertEqual(res.status_code, 400)
+        self.assertIn(
+            "Could not validate value for the text line due to", res.json()["message"]
+        )
+
+        # Missing field value
+        res = self.api_session.post(
+            self.folder_prenotazioni.absolute_url() + "/@booking",
+            json={
+                "booking_date": "%sT09:00:00"
+                % (date.today() + timedelta(2)).strftime("%Y-%m-%d"),
+                "booking_type": "Type A",
+                "fields": [
+                    {"name": "title", "value": "Mario Rossi"},
+                    {"name": "email", "value": "mario.rossi@example"},
+                ],
+                "gate": "Gate A",
+                "additional_fields": [{"name": "text line"}],
+            },
+        )
+        self.assertEqual(res.status_code, 400)
+        self.assertIn(
+            "Additional field 'text line' value is missing.", res.json()["message"]
+        )
+
+        # Text field is not defined
+        res = self.api_session.post(
+            self.folder_prenotazioni.absolute_url() + "/@booking",
+            json={
+                "booking_date": "%sT09:00:00"
+                % (date.today() + timedelta(2)).strftime("%Y-%m-%d"),
+                "booking_type": "Type A",
+                "fields": [
+                    {"name": "title", "value": "Mario Rossi"},
+                    {"name": "email", "value": "mario.rossi@example"},
+                ],
+                "gate": "Gate A",
+                "additional_fields": [{"name": "unexistent", "value": "text"}],
+            },
+        )
+        self.assertEqual(res.status_code, 400)
+        self.assertIn("Unknown additional field 'unexistent'.", res.json()["message"])
 
 
 class TestPrenotazioniIntegrationTesting(unittest.TestCase):